+++ packages/components/config-provider/type.ts
@@ -9,8 +9,9 @@ import { CalendarController } from '../calendar';
 import { ButtonProps } from '../button';
 import { FormErrorMessage } from '../form';
 import { ImageProps } from '../image';
+import { LoadingProps } from '../loading';
 import { MessageOptions } from '../message';
-import { TNode, TElement, SizeEnum, AttachNode } from '../common';
+import type { TNode, TElement, SizeEnum, AttachNode } from '../common';
 
 export interface TdConfigProviderProps {
   /**
@@ -110,6 +111,10 @@ export interface GlobalConfigProvider {
    * 列表组件全局配置
    */
   list?: ListConfig;
+  /**
+   * 加载中组件全局配置
+   */
+  loading?: LoadingConfig;
   /**
    * 消息组件全局配置
    */
@@ -457,6 +462,15 @@ export interface DialogConfig {
    * 确认按钮主题色，即 Dialog 的 `theme` 和 确认按钮的 `theme` 映射关系。示例：{ danger: 'danger' }
    */
   confirmBtnTheme?: { default: string; info: string; warning: string; danger: string; success: string };
+  /**
+   * 全局配置的对话框位置，垂直水平居中显示 和 靠近顶部（top:20%）显示
+   * @default top
+   */
+  placement?: 'top' | 'center';
+  /**
+   * 对话框层级，默认为 2500
+   */
+  zIndex?: number;
 }
 
 export interface DrawerConfig {
@@ -573,15 +587,15 @@ export interface ImageViewerConfig {
    */
   originalSizeTipText?: string;
   /**
-   * 全局语言配置，默认为 “旋转”
-   * @default ''
-   */
-  rotateTipText?: string;
-  /**
    * 全局语言配置，默认为 “预览”
    * @default ''
    */
   previewText?: string;
+  /**
+   * 全局语言配置，默认为 “旋转”
+   * @default ''
+   */
+  rotateTipText?: string;
 }
 
 export interface InputConfig {
@@ -614,6 +628,8 @@ export interface ListConfig {
    */
   loadingText?: string;
 }
+
+export interface LoadingConfig extends LoadingProps {}
 
 export interface MessageConfig extends MessageOptions {}