+++ packages/components/input-number/type.ts
@@ -5,8 +5,8 @@
  * */
 
 import { InputProps } from '../input';
-import { TNode } from '../common';
-import { MouseEvent, KeyboardEvent, FocusEvent, FormEvent, CompositionEvent } from 'react';
+import type { TNode } from '../common';
+import type { MouseEvent, KeyboardEvent, FocusEvent, FormEvent, CompositionEvent } from 'react';
 
 export interface TdInputNumberProps<T = InputNumberValue> {
   /**
@@ -23,6 +23,11 @@ export interface TdInputNumberProps<T = InputNumberValue> {
    * @default false
    */
   autoWidth?: boolean;
+  /**
+   * 自动聚焦
+   * @default false
+   */
+  autofocus?: boolean;
   /**
    * [小数位数](https://en.wiktionary.org/wiki/decimal_place)
    */