+++ packages/components/tag/_example/shape.tsx
@@ -1,4 +1,4 @@
-﻿import React from 'react';
+import React from 'react';
 import { Space, Tag } from 'tdesign-react';
 
 export default function ShapeTagExample() {