+++ script/generate-usage/config.js
@@ -686,7 +686,7 @@ module.exports = {
       const panelList = [{ label: 'list:', value: 'list:' }];
     `,
     usageStr: `
-      const avatarUrl = 'https://tdesign.gtimg.com/list-icon.png';
+      const avatarUrl = 'https://tdesign.gtimg.com/site/images/list-icon.png';
       const listData = [
         { id: 1, content: '列表内容列表内容列表内容' },
         { id: 2, content: '列表内容列表内容列表内容' },